move PermissionsPanel to permissions directory
