+++ src/containers/org/components/index.js
@@ -8,7 +8,6 @@ export { default as AddRoleToOrgModal } from './AddRoleToOrgModal';
 export { default as CreateOrgModal } from './CreateOrgModal';
 export { default as DataSetSearchResultCard } from './DataSetSearchResultCard';
 export { default as EditRoleDetailsBody } from './EditRoleDetailsBody';
-export { default as PermissionsPanel } from './PermissionsPanel';
 export { default as RemoveMemberFromOrgModal } from './RemoveMemberFromOrgModal';
 export { default as RemoveRoleFromMemberModal } from './RemoveRoleFromMemberModal';
 export { default as RenameOrgDatabaseModal } from './RenameOrgDatabaseModal';
