+++ src/containers/permissions/PermissionsPanel.js
@@ -39,10 +39,10 @@ import type {
 } from 'lattice';
 import type { RequestState } from 'redux-reqseq';
 
-import { Divider, SpaceBetweenGrid } from '../../../components';
-import { SET_PERMISSIONS, setPermissions } from '../../../core/permissions/actions';
-import { PERMISSIONS } from '../../../core/redux/constants';
-import { selectDataSetProperties, selectPermissionsByPrincipal } from '../../../core/redux/selectors';
+import { Divider, SpaceBetweenGrid } from '../../components';
+import { SET_PERMISSIONS, setPermissions } from '../../core/permissions/actions';
+import { PERMISSIONS } from '../../core/redux/constants';
+import { selectDataSetProperties, selectPermissionsByPrincipal } from '../../core/redux/selectors';
 
 const { NEUTRAL, PURPLE } = Colors;
 const { APP_CONTENT_PADDING } = Sizes;
