+++ src/containers/permissions/index.js
@@ -5,3 +5,4 @@
 export { default as DataSetPermissionsContainer } from './DataSetPermissionsContainer';
 export { default as ObjectPermissionsContainer } from './ObjectPermissionsContainer';
 export { default as PermissionsActionsGrid } from './PermissionsActionsGrid';
+export { default as PermissionsPanel } from './PermissionsPanel';
